backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
tensorflow
pillow
firebase-admin
requests
sqlmodel
